+++ android/app/src/main/java/dev/cursorremote/android/CursorRemoteApplication.kt
@@ -2,6 +2,7 @@ package dev.cursorremote.android
 
 import android.app.Application
 import dev.cursorremote.android.di.AppContainer
+import dev.cursorremote.android.notify.AppForegroundTracker
 
 class CursorRemoteApplication : Application() {
     lateinit var container: AppContainer
@@ -10,5 +11,10 @@ class CursorRemoteApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         container = AppContainer(this)
+        registerActivityLifecycleCallbacks(
+            AppForegroundTracker { isForeground ->
+                container.pushNotificationCoordinator.setForeground(isForeground)
+            },
+        )
     }
 }